// types/pokemon.ts
export interface PokemonListResponse {
    count: number;
    next: string;
    previous: null;
    results: Result[];
  }
  
  export interface Result {
    name: string;
    url: string;
  }